12.11All

+++ 04.SourceCode/WPFMVVMDemo-master/Operation/BasicOpreation.cs
@@ -11,23 +11,22 @@ namespace Operation
     //二元运算
     public class Basic_Opreation
     {
-        public decimal Add(string num1, string num2)
+        public Double Add(string num1, string num2)
         {
             
-            return Convert.ToDecimal(num1) + Convert.ToDecimal(num2);
+            return Convert.ToDouble(num1) + Convert.ToDouble(num2);
 
         }
-        public decimal Sub(string num1, string num2)
+        public Double Sub(string num1, string num2)
         {
             
-            return Convert.ToDecimal(num1) - Convert.ToDecimal(num2);
+            return Convert.ToDouble(num1) - Convert.ToDouble(num2);
         }
-        public decimal Mul(string num1, string num2)
+        public Double Mul(string num1, string num2)
         {
-            
-            return Convert.ToDecimal(num1) * Convert.ToDecimal(num2);
+            return Convert.ToDouble(num1) * Convert.ToDouble(num2);
         }
-        public decimal Div(string num1, string num2)
+        public Double Div(string num1, string num2)
         {
             
 
@@ -36,7 +35,7 @@ namespace Operation
                 Cache.topCache = "除数不能为零";
 
             }
-                return Convert.ToDecimal(num1) / Convert.ToDecimal(num2);
+                return Convert.ToDouble(num1) / Convert.ToDouble(num2);
 
             
         }
